vector_db/requirements.txt
faiss-cpu==1.9.0.post1
huggingface-hub==0.27.0
numpy==2.2.0
pillow==11.0.0
torch==2.5.1
transformers==4.47.1
flask
